pyproject.toml
[build-system]
requires = ["flit_core >=3.2,<4"]
build-backend = "flit_core.buildapi"

[project]
name = "analogue"
description="ANA/lyze/LOG: Tooling for analyzing Apache access logs"
requires-python = ">=3.10"
authors = [{name = "Robert Grimm", email = "apparebit@gmail.com"}]

classifiers = [
    "Environment :: Console",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.10",
    "Topic :: Internet :: Log Analysis",
    "Topic :: Internet :: WWW/HTTP :: HTTP Servers",
    "Topic :: Utilities",
]

dependencies = [
    "konsole",

    # Parser
    # ------

    "geoip2",  # IP location DB; brew install libmaxminddb
    "ua-parser-up2date",  # User agent properties

    # Analysis
    # --------

    "pandas >= 1.4.3",  # Dataframes
    "pandas-stubs",  # Typing
    "numexpr >= 2.7.1",  # Pandas Performance
    "bottleneck >= 1.3.1",  #_Pandas Performance
    "pyarrow >= 1.0.1",  # Parquet
    "matplotlib >= 3.3.2",  # Graphs
]

dynamic = ["version"]

[project.optional-dependencies]
dev = [
    "black[jupyter]",
    "flit",
    "mypy",
    "pytest",
]

[project.scripts]
analog = "analog.__main__:main"

[project.urls]
Home = "https://github.com/apparebit/analog"

[tool.flit.module]
name = "analog"

[tool.mypy]
python_version = "3.10"
files = ["analog/**/*.py"]
mypy_path = ["stubs"]
warn_unused_configs = true
disallow_any_unimported = true
# disallow_any_expr = true
# disallow_any_decorated = true
# disallow_any_explicit = true
# disallow_any_generics = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
strict_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
# warn_return_any = true
warn_unreachable = true
strict_equality = true
show_error_codes = true

[tool.black]
target-version = ['py310']
skip-string-normalization = true
include = '\.pyi?$'

[pytest]
capture=false
testpaths=["test"]
